requirements.txt
svgwrite
requests
python-dotenv
